requirements.txt
camelot-py[cv]
pandas
requests
Pillow
openpyxl